Other_Python_Projects/discord-ls-parser-bot/requirements.txt
aiohttp~=3.10.10
redis~=5.1.1
discord.py
fake-useragent